Move call test to corresponding module
